Isolating api and views from production code
