+++ frontend/urls.py
@@ -2,7 +2,8 @@ from django.urls import path
 from django.views.generic import TemplateView
 
 from frontend.student_views.angola_estima import ChartBudgetYearView
-from frontend.views import IndexView, CountryView, CountriesExpensesView, ExpensesAndRevenues, TestView, BudgetCountryYear
+from frontend.student_views.angola_lupossa import BudgetCountryYear
+from frontend.views import IndexView, CountryView, CountriesExpensesView, ExpensesAndRevenues, TestView
 
 urlpatterns = [
     path('', IndexView.as_view(), name='index'),
@@ -11,7 +12,7 @@ urlpatterns = [
     path('dataviz/', TemplateView.as_view(template_name="frontend/chart_sample.htm"), name='dataviz'),
     # Students
     path('students/angola/chart-budget-per-year-and-currency/', ChartBudgetYearView.as_view(), name='chart-budget'),
-    path('budget_country_year/', BudgetCountryYear.as_view(), name='budget_country_year'),
+    path('students/angola/budget_country_year/', BudgetCountryYear.as_view(), name='budget_country_year'),
 
     path('<slug>/', CountryView.as_view(), name='country-details'),
     path('<slug>/despesas-e-receitas/', ExpensesAndRevenues.as_view(), name='despesas-e-receitas'),
