+++ frontend/views.py
@@ -6,8 +6,6 @@ from django.views import View
 from django.views.generic.detail import SingleObjectMixin
 
 from apps.geo.models import Country
-from frontend.form import ExerciseAngolaForm
-from frontend.forms import CountryForm
 from frontend.tutorial import COUNTRY_DETAILS_TUTORIAL, INDEX_DIMENSIONS
 
 
@@ -78,13 +76,3 @@ class ExpensesAndRevenues(View, ):
             "country": Country.objects.get(slug=self.kwargs.get("slug"))
         }
         return render(request, 'frontend/expenses-and-revenues.html', context=cn)
-
-
-class BudgetCountryYear(View):
-    def get(self, request):
-
-        form = CountryForm()
-        ctx = {
-            'form': form,
-            }
-        return render(request, 'frontend/budget_country_year.html', context=ctx)
